requirements.txt
PyYAML==3.12   
scipy>=0.19.0    
lxml    
dynet    
matplotlib    
cython   
h5py
asteval
numpy
